Allow compilation with GCC's -mshort option

+++ Source/Data/font.h
@@ -6,7 +6,7 @@
 extern Byte           const * const charmap;
 extern Byte           const * const font_huffman_table;
 extern DWord          const *       font;
-extern unsigned int   const         font_words;
+extern unsigned long  const         font_words;
 
 #define BASE_CHAR 32
