+++ Source/Include/common.h
@@ -17,6 +17,6 @@
 
 typedef unsigned char Byte;
 typedef unsigned short Word;
-typedef unsigned int DWord;
+typedef unsigned long DWord;
 
 #endif
